Extract the MongoDB collection name to a config file
There were a couple of different places where the MongoDB collection name ('trades') was defined, which would require changing it in multiple places if it were to change. This commit defines that variable inside a single config file so it can be imported everywhere else it is needed.

Also, for the sake of more descriptive naming, this commit changes the name of the variable that holds the Bitfinex streaming API endpoint, as well as the filename that holds that variable definition.