Extract rate of return to its own class
Previously, a rate of return was calculated by passing two trades into a method on the Trade class. Now, there is a RateOfReturn class that can be instantiated with two trades, and has an attribute 'rate' which holds the calculated rate of return. This setup allows the Trade class to focus on managing the response to a single trade, while the responsibility for relating many trades together via a rate of return lives in a dedicated class.

Another reason for this change is that we will have to save rates of return to the database (in order to calculate the standard deviation of a collection of rates of return), so it makes sense to have a model that we can instantiate to represent a new rate of return instance in the database.