Add a test for inserting a rate of return into the database
When a rate of return instance is inserted into the database, a Promise should be returned whose resolve value is the newly created document. Also, the count of documents in the database should increase by one.

Also, this commit changes the configuration module containing the names of database collections. The file which exports that module is now called db_collections.js, and it exports two collection names called TRADES_COLLECTION and RATES_OF_RETURN_COLLECTION.