Implement getting all returns in a trailing time period
This commit implements the method RateOfReturn.getTrailing(), which takes an object with the keys 'hours' and 'db' (representing the number of hours in the trailing time period, and the database connection instance). This method returns an Array containing all of the rate of return objects in the database within the specified trailing number of hours.