Add a test for calculating the standard deviation of returns
This commit adds a test for a class Statistics with a static method standardDeviation(ratesOfReturn). This method takes an array of objects which each must have a 'rate' property. It calculates and returns the standard deviation of those rates.

Because this calculation uses floating point arithmetic, the result is not mathematically precise. So, this test only expects the result to be within a tight range, instead of expecting an exact number.