Update unit test for Standard Deviation
This commit updates the unit test for calculating a standard deviation so that the Statistics.standardDeviation(numbers) method accepts an array of numbers as input. This change is in preparation for a different way of calculating price volatility that analyzes the standard deviation of prices themselves, instead of the standard deviation of rates of return between prices.

Also, the test now expects that the standard deviation is rounded to 6 decimal places, which is a somewhat arbitrary level of precision. It seems precise enough to capture small changes in the price volatility, without being so precise that it could get messed up by floating point arithmetic. This level of precision seems reasonable for now.