Update Standard Deviation calculation to use prices
This commit changes the Statistics.standardDeviation(numbers) method to accept an array of numbers as input. For example, these numbers could represent the spot prices of many trades.

This approach enables a change in the way volatility can be calculated. In previous versions of this software, volatility represented the standard deviation of the rates of return between trades at different prices. It turns out that that isn't the only way to represent volatility, and there isn't exactly consensus on what the best approach is. Volatility could also be calculated as the standard deviation of prices themselves, as described in the section 'Calculating Volatility' of this page: https://www.investopedia.com/terms/v/volatility.asp

With the changes in this commit, it will be possible to represent volatility as the standard deviation of a set of prices, which is maybe a bit more intuitive that the standard deviation of the rates of return between prices.