Update price volatility calculation
This commit implements changes that define volatility as the standard deviation of prices over the trailing 5 minutes. This represents a change from the former approach, where volatility was defined as the standard deviation of rates of return between prices over a longer trailing time period of 24 hours.

The new calculation seems more intuitive, and is based on the example calculation provided here: https://www.investopedia.com/terms/v/volatility.asp. The shorter trailing time period also makes the volatility calculation more representative of very short-term changes in market behavior, which might be of interest to traders who are looking for short-term indicators.