Simplify setting up the test database
This commit makes the process of setting up the test database easier and less error-prone.

Previously, a user had to create a test database, configure a database connection string, create a file 'secrets/testDatabaseUri', export the database connection string from that file, and run '$ npm test'.

Now, a user still has to create a test database and configure a database connection string, but doesn't have to create any new files and instead can set up the database connection from the command line at the moment of running the tests. The command to run the tests is: '$ MONGODB_URI=<test database connection string> npm test'. The README contains a full set of instructions for configuring the database connection string and running the tests.