Fix a bug in time calculation
There is a line of code that calculates the timestamp of 5 minutes ago in order to collect all of the trades after that timestamp. A piece of that calculation was using 300000 instead of the correct 300 (maybe from wrongly assuming that the Unix Epoch timestamp format is in milliseconds).

That line of code has been updated to correctly calculate the Unix Epoch timestamp of 5 minutes ago. The Unix Epoch timestamp is in seconds, so 5 minutes ago is actually 300 seconds ago, which is now reflected in the code.