Clean stale records from the database
Previously, the app was accumulating records in the database without ever deleting any records. This commit sets up the app to delete all records from earlier than the beginning of the analysis period of the trailing 5 minutes. This cleanup process happens every time a new record is entered into the database.