Display the price volatility in a web ui
This commit creates state in the React application with a 'volatility' key whose value is an integer to represent the price volatility. When the WebSocket server emits a message that contains a price volatility value, the React app calls a handler that re-sets the app state to incorporate the new volatility value. React updates the UI to display this value.

Currently, the React app listens to a local WebSocket server at ws://localhost:3000, which will eventually need to be changed to a production URL.