Listen for WebSocket connections on port 8080
This commit sets up the server to listen for WebSocket connections on port 8080. This port number is arbitrary, but there appears to be precedent for using 8080 in cases where there isn't a standard service port that corresponds to the type of traffic needed to access the application.

In order for this server to successfully run on an EC2 instance in production, a security group needs to be configured to allow TCP traffic over port 8080.