Move README back to the root directory
In an earlier commit where this repository was set up to house both server-side and client-side code, the README stayed in the server directory, which caused it to be hidden from the GitHub web user interface. This commit moves the README back into the repository's root directory so that GitHub can display it on the repository's web page again.